session completed

+++ DJ/todo/views.py
@@ -7,6 +7,11 @@ from .models import uc
 
 # Create your views here.
 def home(req):
+    print(dict(req.session))
+    if( not dict(req.session)):
+        return redirect('login')
+
+
     return render(req,'home.html',{"name":"python"})
 
 users = [
@@ -17,7 +22,8 @@ users = [
         {"name":"flutter","age":20,"email":"flutter@gmail.com"},
         ]
 def about(req):
-   
+    if( not dict(req.session)):
+        return redirect('login')
     return render(req,'about.html',{"isDone":'z',"users":users})
 
 def work(req):
@@ -38,6 +44,8 @@ def work(req):
 
 
 def viewUser(req):
+    if( not dict(req.session)):
+        return redirect('login')
     dbData = list(uc.find())
     datas = []
     for i in dbData:
@@ -50,6 +58,45 @@ def viewUser(req):
 
 
 def deleteUser(req,id):
+    if( not dict(req.session)):
+        return redirect('login')
     print(id,"docId for delete operation")
     uc.delete_one({"_id":ObjectId(id)})
     return redirect("info")
+
+
+def updateUser(req,id):
+    if( not dict(req.session)):
+        return redirect('login')
+    userData = uc.find_one({"_id":ObjectId(id)})
+    if(req.method == 'POST'):
+        username = req.POST['username']
+        email = req.POST['email']
+        password = req.POST['password']
+        x = datetime.datetime.now()
+       
+        data = {"username":username,"email":email,"password":password,"updated_time":x.strftime("%d-%B-%y %X")}
+        print(data)
+        fltr = {"_id":ObjectId(id)}
+        uc.update_one(fltr,{"$set":data})
+        return redirect('info')
+    return render(req,'work.html',{"user":userData})
+
+
+def login(req):
+    if(req.method == "POST"):
+        email = req.POST['email']
+        user = uc.find_one({"email":email})
+        print(user,'===========))))))))))')
+        if(user):
+            req.session['docId'] = str(user['_id'])
+            return redirect('home')
+        else:
+            return redirect('work')
+    return render(req,'login.html')
+
+
+
+def logout(req):
+    req.session.clear()
+    return redirect('home')
